requirements.txt
opencv-python==4.9.0.80
face-recognition==1.3.0
pandas==2.2.1
openpyxl==3.1.2
pyserial==3.5
gspread==6.0.2
oauth2client==4.1.3
pillow==10.2.0
numpy==1.26.4 